requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
httpx==0.24.1
python-dotenv==1.0.0
langchain==0.0.287
langgraph==0.0.10
google-generativeai==0.3.0
transformers==4.35.0
torch==2.0.1
pillow==10.0.0
python-multipart==0.0.6
